+++ NEWS.md
@@ -1,8 +1,8 @@
-# 0.4.5
+# 0.5
 - Update LinkedIn API version
 - Use templates
 - Use R Universe for packages
-- Move Usecases etc. workflows to promoutils
+- Move Usecases and help wanted etc. workflows to promoutils
 - Use cli
 - Add consistency
 - Add Throwback Thursday functions